Rename folder which contains root component for runs table.
